backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sse-starlette==1.8.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain & LangGraph
langgraph==0.0.40
langchain==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.1.0
langchain-community==0.0.20

# LlamaIndex
llama-index==0.10.0
llama-index-vector-stores-postgres==0.1.0
llama-index-embeddings-openai==0.1.0
llama-index-llms-openai==0.1.0

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4
supabase==2.3.0
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
tenacity==8.2.3

# Document Processing
pypdf==4.0.0
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
markdown==3.5.2
beautifulsoup4==4.12.3
lxml==5.1.0

# Data Connectors
boto3==1.34.34
botocore==1.34.34
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
Office365-REST-Python-Client==2.5.3
atlassian-python-api==3.41.4

# Monitoring & Logging
langsmith==0.0.87
sentry-sdk[fastapi]==1.40.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
mypy==1.8.0
